+++ Assets/Scripts/BearCollider.cs
