+++ Assets/Scripts/DollarCollider.cs
