+++ Assets/Scripts/HighScoreManager.cs
