+++ Assets/Sounds/StartSound.mp3.meta
