--- Assets/db/readme.md
@@ -1,22 +1,0 @@
-Dla osób które potrzebują trzymania za rączke :) 
-(Przepraszam że nie po angielskiemu ale i don't spik americano)
-1. Proszę pobrac najnowsze zmiany
-2. Żartowałem już je masz skoro to czytasz
-3. Teraz przejdzmy do trudniejszych rzeczy wejdź tu
-https://sqlitebrowser.org/dl/#windows
-Jak masz linuxa to tu:
-https://github.com/sqlitebrowser/sqlitebrowser
-Jeśli nie masz przeglądarki polecam https://www.google.com/intl/pl_pl/chrome/
-4. Jak Ci się uda pobrać to zainstaluj, trzeba klikać "next"
-5. Jeśli przeszedłeś ten krok to już jesteś kozak
-6. Odpal sobie ten program daj otwórz baze danych i wybierz plik z repozytorium assets\db\db.db
-7. No i masz przykładową bazę na której możesz puszczać selecty i się bawić
-Teraz jak to wygląda w unity
-1. Podpiąłem skrypt pod kliknięcie w puchar
-2. Przykładowy kod wyciąga dane z tabeli SCORES(id_score, score)
-3. Dane z tej tabeli pokazują się w debug logu
-4. Jeśli Ci się pokazały to jesteś przekoks
-
-
-P.S 
-Jako iż profesjonalistom nie trzeba tłumaczyć to podejrzewam że kod dbConn.cs jest na tyle prosty że łatwo Wam będzie dodać inserty, updatey i inne takie.
